Support for XSS protection using origin isolation
When using a redirect_uri hosted on a domain other than the one
used by the application code, the access token will be inaccessible
to the application code (due to normal origin isolation behavior).

Network requests to resource server endpoints will be handled by
the identity proxy, which will use message passing to redirect the
request to an iframe hosted in the same origin as the redirect_uri.

In the case of a XSS attack, the malicious code will be unable to
ex-filtrate the access token.