Service Worker thread keepAlive
This important fix addresses an issue whereby the "service worker" identity proxy implementation was failing after a period of inactivity in the browser. The issue was that browsers (most obviously, Firefox) would kill the server worker thread when it hadn't received any messages after a while. When the next message did arrive, the service worker would start back up; unfortunately, when starting back up the references it had been originally passed were lost. This caused the requests to RS endpoints to pass through without the access token, ultimately breaking the app.

The solution is simple - introduce a periodic "keepAlive" message from the parent context to the service worker. So long as the parent context is active, the service worker will also remain active.